requirements.txt
dtaidistance==2.3.13
dtw-python==1.5.3
matplotlib==3.9.2
networkx==2.8.8
numpy==2.1.3
pandas==2.2.3
POT==0.9.5
scikit-learn==1.7.1
scipy==1.14.1
seaborn==0.13.2
shortuuid==1.0.13
tqdm==4.67.1